small frontend refactor, new husky check for types
