+++ frontend/.husky/pre-push
@@ -1,3 +1,4 @@
 cd frontend
 
 npm run lint
+npm run types:check
