+++ frontend/package.json
@@ -16,7 +16,8 @@
     "build:dev": "webpack --env mode=development",
     "build:prod": "webpack --env mode=production",
     "health": "echo \"👀 ----- Test line for testing -----👀\" && exit 0",
-    "codegen": "npx @rtk-query/codegen-openapi openapi-config.ts && eslint . --fix"
+    "codegen": "npx @rtk-query/codegen-openapi openapi-config.ts && eslint . --fix",
+    "types:check": "tsc --noEmit --pretty"
   },
   "repository": {
     "type": "git",
