+++ frontend/src/shared/auth/authSlice.ts
@@ -1,22 +1,8 @@
 import { createSlice } from "@reduxjs/toolkit";
 
 import { UserPayloadUser } from "@/shared/api/generated";
-import z from "zod";
 
-export const authSliceSchema = z.object({
-  isAuthenticated: z.boolean(),
-  user: z
-    .object({
-      username: z.string().nonempty(),
-      email: z.email(),
-      image: z.string().nullable(),
-      bio: z.string(),
-      token: z.jwt(),
-    })
-    .nullable(),
-});
-
-export type AuthSliceState = z.infer<typeof authSliceSchema>;
+import { AuthSliceState } from "./authSlice.schema";
 
 export const authSliceLoggedOutState: AuthSliceState = {
   user: null,
